SVT streamer can now accept requests for fractional mip levels. Smoothly changing the requested fractional mip level can currently lead to certain performance issues with DDC because additional small requests are made to stream in a few missing tiles at a time. To mitigate this, there is now a new cvar to discretize the computed streaming size. Also fixed a crash when unregistering a SVT from the streamer while it is currently playing back. Fixed by properly removing all FStreamingInstances belonging to the SVT to be removed.
#rb Patrick.Kelly
#rnx

[CL 32327980 by tim doerries in ue5-main branch]

+++ Engine/Source/Runtime/Engine/Classes/SparseVolumeTexture/SparseVolumeTexture.h
@@ -372,7 +372,7 @@ public:
 	// If bBlocking is true, DDC streaming requests will block on completion, guaranteeing that the requested frame will have been streamed in after the next streaming system update.
 	// If streaming cooked data from disk, the highest priority will be used, but no guarantee is given.
 	// if bHasValidFrameRate is true, the FrameRate argument will be use to predict the required streaming IO bandwidth.
-	static ENGINE_API USparseVolumeTextureFrame* GetFrameAndIssueStreamingRequest(USparseVolumeTexture* SparseVolumeTexture, uint32 StreamingInstanceKey, float FrameRate, float FrameIndex, int32 MipLevel, bool bBlocking, bool bHasValidFrameRate);
+	static ENGINE_API USparseVolumeTextureFrame* GetFrameAndIssueStreamingRequest(USparseVolumeTexture* SparseVolumeTexture, uint32 StreamingInstanceKey, float FrameRate, float FrameIndex, float MipLevel, bool bBlocking, bool bHasValidFrameRate);
 
 	ENGINE_API bool Initialize(USparseVolumeTexture* InOwner, int32 InFrameIndex, const FTransform& InFrameTransform, UE::SVT::FTextureData& UncookedFrame);
 	int32 GetFrameIndex() const { return FrameIndex; }
